+++ java/ham_fisted/Reducible.java
@@ -1,10 +1,16 @@
 package ham_fisted;
 
 
-import java.util.Collection;
+import java.util.Iterator;
 
 
 public interface Reducible
 {
-  Reducible reduce(Collection<Reducible> rest);
+  default Reducible reduceIter(Iterator<Reducible> rest) {
+    Reducible retval = this;
+    while(rest.hasNext())
+      retval = retval.reduce(rest.next());
+    return retval;
+  }
+  Reducible reduce(Reducible rhs);
 }