+++ test/ham_fisted/api_test.clj
@@ -1,6 +1,7 @@
 (ns ham-fisted.api-test
   (:require [clojure.test :refer [deftest is]]
-            [ham-fisted.api :as hamf]))
+            [ham-fisted.api :as hamf]
+            [ham-fisted.lazy-noncaching :as lznc]))
 
 
 
@@ -18,3 +19,27 @@
   (is (thrown? Exception (hamf/pgroups (fn [^long sidx ^long eidx]
                                          (when (>= sidx 70)
                                            (throw (Exception. "Error!!"))) sidx)))))
+
+
+(deftest group-by-nil
+  (is (= {} (hamf/group-by :a nil)))
+  (is (= {} (hamf/group-by-reduce :a + nil nil)))
+  (is (= {} (hamf/group-by :a {})))
+  (is (= {} (hamf/group-by-reduce :a + nil {})))
+  )
+
+
+(deftest map-filter-concat-nil
+  (is (= (map + nil) (lznc/map + nil)))
+  (is (= (filter + nil) (lznc/filter + nil)))
+  (is (= (concat) (lznc/concat)))
+  (is (= (concat nil) (lznc/concat nil))))
+
+
+(deftest conj-with-friends
+  (is (= (conj (map inc (list 1 2 3 4)) 4)
+         (conj (lznc/map inc (list 1 2 3 4)) 4)))
+  (is (= (conj (filter even? (list 1 2 3 4)) 4)
+         (conj (lznc/filter even? (list 1 2 3 4)) 4)))
+  (is (= (conj (concat [1 2 3] [4 5 6]) 4)
+         (conj (lznc/concat [1 2 3] [4 5 6]) 4))))